+++ src/Listener/ConfigListener.php
@@ -14,6 +14,7 @@ use Zend\Config\Config;
 use Zend\Config\Factory as ConfigFactory;
 use Zend\EventManager\EventManagerInterface;
 use Zend\EventManager\ListenerAggregateInterface;
+use Zend\EventManager\ListenerAggregateTrait;
 use Zend\ModuleManager\Feature\ConfigProviderInterface;
 use Zend\ModuleManager\ModuleEvent;
 use Zend\Stdlib\ArrayUtils;
@@ -26,13 +27,10 @@ class ConfigListener extends AbstractListener implements
     ConfigMergerInterface,
     ListenerAggregateInterface
 {
+    use ListenerAggregateTrait;
+
     const STATIC_PATH = 'static_path';
     const GLOB_PATH   = 'glob_path';
-
-    /**
-     * @var array
-     */
-    protected $callbacks = [];
 
     /**
      * @var array
@@ -79,18 +77,18 @@ class ConfigListener extends AbstractListener implements
     /**
      * {@inheritDoc}
      */
-    public function attach(EventManagerInterface $events)
+    public function attach(EventManagerInterface $events, $priority = 1)
     {
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onloadModulesPre'], 1000);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onloadModulesPre'], 1000);
 
         if ($this->skipConfig) {
             // We already have the config from cache, no need to collect or merge.
             return $this;
         }
 
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, [$this, 'onLoadModule']);
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onLoadModules'], -1000);
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_MERGE_CONFIG, [$this, 'onMergeConfig'], 1000);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, [$this, 'onLoadModule']);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onLoadModules'], -1000);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_MERGE_CONFIG, [$this, 'onMergeConfig'], 1000);
         return $this;
     }
 
@@ -165,11 +163,15 @@ class ConfigListener extends AbstractListener implements
     {
         // Trigger MERGE_CONFIG event. This is a hook to allow the merged application config to be
         // modified before it is cached (In particular, allows the removal of config keys)
-        $e->getTarget()->getEventManager()->trigger(ModuleEvent::EVENT_MERGE_CONFIG, $e);
+        $originalEventName = $e->getName();
+        $e->setName(ModuleEvent::EVENT_MERGE_CONFIG);
+        $e->getTarget()->getEventManager()->triggerEvent($e);
+
+        // Reset event name
+        $e->setName($originalEventName);
 
         // If enabled, update the config cache
-        if (
-            $this->getOptions()->getConfigCacheEnabled()
+        if ($this->getOptions()->getConfigCacheEnabled()
             && false === $this->skipConfig
         ) {
             $configFile = $this->getOptions()->getConfigCacheFile();
@@ -177,18 +179,6 @@ class ConfigListener extends AbstractListener implements
         }
 
         return $this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function detach(EventManagerInterface $events)
-    {
-        foreach ($this->callbacks as $index => $callback) {
-            if ($events->detach($callback)) {
-                unset($this->callbacks[$index]);
-            }
-        }
     }
 
     /**