+++ src/Listener/DefaultListenerAggregate.php
@@ -12,7 +12,6 @@ namespace Zend\ModuleManager\Listener;
 use Zend\EventManager\EventManagerInterface;
 use Zend\EventManager\ListenerAggregateInterface;
 use Zend\ModuleManager\ModuleEvent;
-use Zend\Stdlib\CallbackHandler;
 
 /**
  * Default listener aggregate
@@ -34,18 +33,24 @@ class DefaultListenerAggregate extends AbstractListener implements
      * Attach one or more listeners
      *
      * @param  EventManagerInterface $events
+     * @param  int $priority
      * @return DefaultListenerAggregate
      */
-    public function attach(EventManagerInterface $events)
+    public function attach(EventManagerInterface $events, $priority = 1)
     {
         $options                     = $this->getOptions();
         $configListener              = $this->getConfigListener();
         $locatorRegistrationListener = new LocatorRegistrationListener($options);
+        $moduleLoaderListener        = new ModuleLoaderListener($options);
 
-        // High priority, we assume module autoloading (for FooNamespace\Module classes) should be available before anything else
-        $this->listeners[] = $events->attach(new ModuleLoaderListener($options));
+        // High priority, we assume module autoloading (for FooNamespace\Module
+        // classes) should be available before anything else
+        $moduleLoaderListener->attach($events);
+        $this->listeners[] = $moduleLoaderListener;
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE_RESOLVE, new ModuleResolverListener);
-        // High priority, because most other loadModule listeners will assume the module's classes are available via autoloading
+
+        // High priority, because most other loadModule listeners will assume
+        // the module's classes are available via autoloading
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, new AutoloaderListener($options), 9000);
 
         if ($options->getCheckDependencies()) {
@@ -54,8 +59,11 @@ class DefaultListenerAggregate extends AbstractListener implements
 
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, new InitTrigger($options));
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, new OnBootstrapListener($options));
-        $this->listeners[] = $events->attach($locatorRegistrationListener);
-        $this->listeners[] = $events->attach($configListener);
+
+        $locatorRegistrationListener->attach($events);
+        $configListener->attach($events);
+        $this->listeners[] = $locatorRegistrationListener;
+        $this->listeners[] = $configListener;
         return $this;
     }
 
@@ -68,19 +76,14 @@ class DefaultListenerAggregate extends AbstractListener implements
     public function detach(EventManagerInterface $events)
     {
         foreach ($this->listeners as $key => $listener) {
-            $detached = false;
-            if ($listener === $this) {
+            if ($listener instanceof ListenerAggregateInterface) {
+                $listener->detach($events);
+                unset($this->listeners[$key]);
                 continue;
             }
-            if ($listener instanceof ListenerAggregateInterface) {
-                $detached = $listener->detach($events);
-            } elseif ($listener instanceof CallbackHandler) {
-                $detached = $events->detach($listener);
-            }
 
-            if ($detached) {
-                unset($this->listeners[$key]);
-            }
+            $events->detach($listener);
+            unset($this->listeners[$key]);
         }
     }