+++ src/Listener/LocatorRegistrationListener.php
@@ -11,6 +11,7 @@ namespace Zend\ModuleManager\Listener;
 
 use Zend\EventManager\EventManagerInterface;
 use Zend\EventManager\ListenerAggregateInterface;
+use Zend\EventManager\ListenerAggregateTrait;
 use Zend\ModuleManager\Feature\LocatorRegisteredInterface;
 use Zend\ModuleManager\ModuleEvent;
 use Zend\ModuleManager\ModuleManager;
@@ -22,15 +23,12 @@ use Zend\Mvc\MvcEvent;
 class LocatorRegistrationListener extends AbstractListener implements
     ListenerAggregateInterface
 {
-    /**
-     * @var array
-     */
-    protected $modules = [];
+    use ListenerAggregateTrait;
 
     /**
      * @var array
      */
-    protected $callbacks = [];
+    protected $modules = [];
 
     /**
      * loadModule
@@ -43,7 +41,7 @@ class LocatorRegistrationListener extends AbstractListener implements
      */
     public function onLoadModule(ModuleEvent $e)
     {
-        if (!$e->getModule() instanceof LocatorRegisteredInterface) {
+        if (! $e->getModule() instanceof LocatorRegisteredInterface) {
             return;
         }
         $this->modules[] = $e->getModule();
@@ -62,18 +60,18 @@ class LocatorRegistrationListener extends AbstractListener implements
         $moduleManager = $e->getTarget();
         $events        = $moduleManager->getEventManager()->getSharedManager();
 
-        if (!$events) {
+        if (! $events) {
             return;
         }
 
         // Shared instance for module manager
         $events->attach('Zend\Mvc\Application', ModuleManager::EVENT_BOOTSTRAP, function (MvcEvent $e) use ($moduleManager) {
-            $moduleClassName = get_class($moduleManager);
+            $moduleClassName      = get_class($moduleManager);
             $moduleClassNameArray = explode('\\', $moduleClassName);
             $moduleClassNameAlias = end($moduleClassNameArray);
-            $application     = $e->getApplication();
-            $services        = $application->getServiceManager();
-            if (!$services->has($moduleClassName)) {
+            $application          = $e->getApplication();
+            $services             = $application->getServiceManager();
+            if (! $services->has($moduleClassName)) {
                 $services->setAlias($moduleClassName, $moduleClassNameAlias);
             }
         }, 1000);
@@ -104,7 +102,7 @@ class LocatorRegistrationListener extends AbstractListener implements
 
         foreach ($this->modules as $module) {
             $moduleClassName = get_class($module);
-            if (!$services->has($moduleClassName)) {
+            if (! $services->has($moduleClassName)) {
                 $services->setService($moduleClassName, $module);
             }
         }
@@ -113,22 +111,10 @@ class LocatorRegistrationListener extends AbstractListener implements
     /**
      * {@inheritDoc}
      */
-    public function attach(EventManagerInterface $events)
+    public function attach(EventManagerInterface $events, $priority = 1)
     {
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, [$this, 'onLoadModule']);
-        $this->callbacks[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onLoadModules'], -1000);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, [$this, 'onLoadModule']);
+        $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES, [$this, 'onLoadModules'], -1000);
         return $this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function detach(EventManagerInterface $events)
-    {
-        foreach ($this->callbacks as $index => $callback) {
-            if ($events->detach($callback)) {
-                unset($this->callbacks[$index]);
-            }
-        }
     }
 }