+++ src/Listener/ServiceListener.php
@@ -13,11 +13,19 @@ use Traversable;
 use Zend\EventManager\EventManagerInterface;
 use Zend\ModuleManager\ModuleEvent;
 use Zend\ServiceManager\Config as ServiceConfig;
+use Zend\ServiceManager\ConfigInterface as ServiceConfigInterface;
 use Zend\ServiceManager\ServiceManager;
 use Zend\Stdlib\ArrayUtils;
 
 class ServiceListener implements ServiceListenerInterface
 {
+    /**
+     * Service manager post-configuration.
+     *
+     * @var ServiceManager
+     */
+    protected $configuredServiceManager;
+
     /**
      * @var \Zend\Stdlib\CallbackHandler[]
      */
@@ -31,6 +39,8 @@ class ServiceListener implements ServiceListenerInterface
     protected $defaultServiceManager;
 
     /**
+     * Default service configuration for the application service manager.
+     *
      * @var array
      */
     protected $defaultServiceConfig;
@@ -59,18 +69,12 @@ class ServiceListener implements ServiceListenerInterface
      */
     public function setDefaultServiceConfig($configuration)
     {
-        $this->defaultServiceConfig  = $configuration;
-
+        $this->defaultServiceConfig = $configuration;
         return $this;
     }
 
     /**
-     * @param  ServiceManager|string $serviceManager  Service Manager instance or name
-     * @param  string                $key             Configuration key
-     * @param  string                $moduleInterface FQCN as string
-     * @param  string                $method          Method name
-     * @throws Exception\RuntimeException
-     * @return ServiceListener
+     * @inheritDoc
      */
     public function addServiceManager($serviceManager, $key, $moduleInterface, $method)
     {
@@ -81,7 +85,7 @@ class ServiceListener implements ServiceListenerInterface
         } else {
             throw new Exception\RuntimeException(sprintf(
                 'Invalid service manager provided, expected ServiceManager or string, %s provided',
-                (string) $serviceManager
+                (is_object($serviceManager) ? get_class($serviceManager) : gettype($serviceManager))
             ));
         }
 
@@ -102,9 +106,10 @@ class ServiceListener implements ServiceListenerInterface
 
     /**
      * @param  EventManagerInterface $events
+     * @param  int $priority
      * @return ServiceListener
      */
-    public function attach(EventManagerInterface $events)
+    public function attach(EventManagerInterface $events, $priority = 1)
     {
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULE, [$this, 'onLoadModule']);
         $this->listeners[] = $events->attach(ModuleEvent::EVENT_LOAD_MODULES_POST, [$this, 'onLoadModulesPost']);
@@ -144,15 +149,15 @@ class ServiceListener implements ServiceListenerInterface
         $module = $e->getModule();
 
         foreach ($this->serviceManagers as $key => $sm) {
-            if (!$module instanceof $sm['module_class_interface']
-                && !method_exists($module, $sm['module_class_method'])
+            if (! $module instanceof $sm['module_class_interface']
+                && ! method_exists($module, $sm['module_class_method'])
             ) {
                 continue;
             }
 
             $config = $module->{$sm['module_class_method']}();
 
-            if ($config instanceof ServiceConfig) {
+            if ($config instanceof ServiceConfigInterface) {
                 $config = $this->serviceConfigToArray($config);
             }
 
@@ -160,12 +165,12 @@ class ServiceListener implements ServiceListenerInterface
                 $config = ArrayUtils::iteratorToArray($config);
             }
 
-            if (!is_array($config)) {
-                // If we don't have an array by this point, nothing left to do.
+            if (! is_array($config)) {
+                // If we do not have an array by this point, nothing left to do.
                 continue;
             }
 
-            // We're keeping track of which modules provided which configuration to which service managers.
+            // We are keeping track of which modules provided which configuration to which service managers.
             // The actual merging takes place later. Doing it this way will enable us to provide more powerful
             // debugging tools for showing which modules overrode what.
             $fullname = $e->getModuleName() . '::' . $sm['module_class_method'] . '()';
@@ -190,34 +195,27 @@ class ServiceListener implements ServiceListenerInterface
         $config         = $configListener->getMergedConfig(false);
 
         foreach ($this->serviceManagers as $key => $sm) {
-            if (isset($config[$sm['config_key']])
-                && is_array($config[$sm['config_key']])
-                && !empty($config[$sm['config_key']])
-            ) {
-                $this->serviceManagers[$key]['configuration']['merged_config'] = $config[$sm['config_key']];
-            }
+            $smConfig = $this->mergeServiceConfiguration($key, $sm, $config);
 
-            // Merge all of the things!
-            $smConfig = [];
-            foreach ($this->serviceManagers[$key]['configuration'] as $configs) {
-                if (isset($configs['configuration_classes'])) {
-                    foreach ($configs['configuration_classes'] as $class) {
-                        $configs = ArrayUtils::merge($configs, $this->serviceConfigToArray($class));
-                    }
-                }
-                $smConfig = ArrayUtils::merge($smConfig, $configs);
-            }
-
-            if (!$sm['service_manager'] instanceof ServiceManager) {
-                $instance = $this->defaultServiceManager->get($sm['service_manager']);
-                if (!$instance instanceof ServiceManager) {
+            if (! $sm['service_manager'] instanceof ServiceManager) {
+                if (! $this->defaultServiceManager->has($sm['service_manager'])) {
                     throw new Exception\RuntimeException(sprintf(
                         'Could not find a valid ServiceManager for %s',
                         $sm['service_manager']
                     ));
                 }
+
+                $instance = $this->defaultServiceManager->get($sm['service_manager']);
+                if (! $instance instanceof ServiceManager) {
+                    throw new Exception\RuntimeException(sprintf(
+                        'Could not find a valid ServiceManager for %s',
+                        $sm['service_manager']
+                    ));
+                }
+
                 $sm['service_manager'] = $instance;
             }
+
             $serviceConfig = new ServiceConfig($smConfig);
             $serviceConfig->configureServiceManager($sm['service_manager']);
         }
@@ -226,12 +224,16 @@ class ServiceListener implements ServiceListenerInterface
     /**
      * Merge a service configuration container
      *
-     * Extracts the various service configuration arrays, and then merges with
-     * the internal service configuration.
+     * Extracts the various service configuration arrays.
      *
-     * @param  ServiceConfig|string $config Instance of ServiceConfig or class name
-     * @throws Exception\RuntimeException
+     * @param ServiceConfigInterface|string $config ServiceConfigInterface or
+     *     class name resolving to one.
      * @return array
+     * @throws Exception\RuntimeException if resolved class name is not a
+     *     ServiceConfigInterface implementation.
+     * @throws Exception\RuntimeException under zend-servicemanager v2 if the
+     *     configuration instance is not specifically a ServiceConfig, as there
+     *     is no way to extract service configuration in that case.
      */
     protected function serviceConfigToArray($config)
     {
@@ -240,21 +242,70 @@ class ServiceListener implements ServiceListenerInterface
             $config = new $class;
         }
 
-        if (!$config instanceof ServiceConfig) {
+        if (! $config instanceof ServiceConfigInterface) {
             throw new Exception\RuntimeException(sprintf(
-                'Invalid service manager configuration class provided; received "%s", expected an instance of Zend\ServiceManager\Config',
-                (is_object($config) ? get_class($config) : (is_scalar($config) ? $config : gettype($config)))
+                'Invalid service manager configuration class provided; received "%s", expected an instance of %s',
+                (is_object($config) ? get_class($config) : (is_scalar($config) ? $config : gettype($config))),
+                ServiceConfigInterface::class
             ));
         }
 
+        if (method_exists($config, 'toArray')) {
+            // zend-servicemanager v3 interface
+            return $config->toArray();
+        }
+
+        // For zend-servicemanager v2, we need a Zend\ServiceManager\Config
+        // instance specifically.
+        if (! $config instanceof ServiceConfig) {
+            throw new Exception\RuntimeException(sprintf(
+                'Invalid service manager configuration class provided; received "%s", expected an instance of %s',
+                (is_object($config) ? get_class($config) : (is_scalar($config) ? $config : gettype($config))),
+                ServiceConfig::class
+            ));
+        }
+
+        // Pull service configuration from discrete methods.
         return [
             'abstract_factories' => $config->getAbstractFactories(),
             'aliases'            => $config->getAliases(),
-            'initializers'       => $config->getInitializers(),
+            'delegators'         => $config->getDelegators(),
             'factories'          => $config->getFactories(),
+            'initializers'       => $config->getInitializers(),
             'invokables'         => $config->getInvokables(),
             'services'           => $config->getServices(),
             'shared'             => $config->getShared(),
         ];
+    }
+
+    /**
+     * Merge all configuration for a given service manager to a single array.
+     *
+     * @param string $key Named service manager
+     * @param array $metadata Service manager metadata
+     * @param array $config Merged configuration
+     * @return array Service manager-specific configuration
+     */
+    private function mergeServiceConfiguration($key, array $metadata, array $config)
+    {
+        if (isset($config[$metadata['config_key']])
+            && is_array($config[$metadata['config_key']])
+            && !empty($config[$metadata['config_key']])
+        ) {
+            $this->serviceManagers[$key]['configuration']['merged_config'] = $config[$metadata['config_key']];
+        }
+
+        // Merge all of the things!
+        $serviceConfig = [];
+        foreach ($this->serviceManagers[$key]['configuration'] as $name => $configs) {
+            if (isset($configs['configuration_classes'])) {
+                foreach ($configs['configuration_classes'] as $class) {
+                    $configs = ArrayUtils::merge($configs, $this->serviceConfigToArray($class));
+                }
+            }
+            $serviceConfig = ArrayUtils::merge($serviceConfig, $configs);
+        }
+
+        return $serviceConfig;
     }
 }