+++ test/Listener/TestAsset/CustomPluginManager.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ModuleManager\Listener\TestAsset;
+
+use Zend\ServiceManager\AbstractPluginManager;
+use Zend\ServiceManager\Exception\InvalidServiceException;
+
+class CustomPluginManager extends AbstractPluginManager
+{
+    protected $instanceOf = CustomPluginInterface::class;
+
+    public function validate($plugin)
+    {
+        if (! $plugin instanceof $this->instanceOf) {
+            throw new InvalidServiceException();
+        }
+    }
+
+    public function validatePlugin($plugin)
+    {
+        $this->validate($plugin);
+    }
+}