+++ test/TestAsset/ListenerTestModule/Module.php
@@ -10,14 +10,14 @@
 namespace ListenerTestModule;
 
 use Zend\ModuleManager\Feature\AutoloaderProviderInterface;
-use Zend\ModuleManager\Feature\LocatorRegisteredInterface;
 use Zend\ModuleManager\Feature\BootstrapListenerInterface;
+use Zend\ModuleManager\Feature\LocatorRegisteredInterface;
 use Zend\EventManager\EventInterface;
 
 class Module implements
     AutoloaderProviderInterface,
-    LocatorRegisteredInterface,
-    BootstrapListenerInterface
+    BootstrapListenerInterface,
+    LocatorRegisteredInterface
 {
     public $initCalled = false;
     public $getConfigCalled = false;